Added ReadTrans in order to read transactions. Also added a check in the CashAccount constructor so when there are no transactions passed in there is an empty array list created.
Transactions and Cash accounts are associated to eachother in "fillUsers" method in the User class. If a transaction has a cash account name that represents a cash account that no longer exists then it will not have an associated cash account object. If a cash account has no transactions it will have an empty array list represent its transactions.